Consolidate timestamp tests
